Added video streaming mechanism
Created a basic react component that, given a video element reference,
will start to capture the images at a given interval and send to the
backend through a websocket endpoint.

Uses json to send in the metadata (for now, a millisecond timestamp
and a mime type is the metadata) along with the image in form of a 80%
compressed jpeg file appended to the json as a direct binary blob.

The backend will have to figure out when the json ends and the binary
data begins.

The UI controls for now is just a few buttons. The result is logged
into the console.