Added some more stream interactions in demo
Now in demo, one can send some additional commands to the streaming
server.

There is a reset command that doesnot have to stop and then start the
streaming process again, and will just reset / flush all the states in
the streaming server background

There is a button that will send a query to get the number of clips.

There is also another button and a input field that is used to set the
nth clip as the backend video for stateless / non-streaming mode